Code refactoring and animations
