+++ package.json
@@ -19,6 +19,7 @@
     "firebase": "^10.7.2",
     "firebaseui": "^6.1.0",
     "framer-motion": "^11.0.3",
+    "lottie-react": "^2.4.0",
     "postcss": "^8.4.38",
     "rc-slider": "^10.5.0",
     "react": "^18.2.0",
